+++ tests/testthat/test-bind.R
@@ -24,13 +24,28 @@ test_that("vec_rbind() output is tibble if any input is tibble", {
 test_that("type of column is common type of individual columns", {
   x_int <- data_frame(x = 1L)
   x_dbl <- data_frame(x = 2.5)
-  x_chr <- data_frame(x = "a")
 
   expect_equal(vec_rbind(x_int, x_int), data_frame(x = c(1L, 1L)))
   expect_equal(vec_rbind(x_int, x_dbl), data_frame(x = c(1, 2.5)))
+})
+
+test_that("incompatible columns throws common type error", {
+  x_int <- data_frame(x = 1L)
+  x_chr <- data_frame(x = "a")
 
   expect_snapshot({
-    (expect_error(vec_rbind(x_int, x_chr), class = "vctrs_error_incompatible_type"))
+    (expect_error(
+      vec_rbind(x_int, x_chr),
+      class = "vctrs_error_incompatible_type"
+    ))
+    (expect_error(
+      vec_rbind(x_int, x_chr, .call = call("foo")),
+      class = "vctrs_error_incompatible_type"
+    ))
+    (expect_error(
+      vec_rbind(x_int, x_chr, .ptype = x_chr, .call = call("foo")),
+      class = "vctrs_error_incompatible_type"
+    ))
   })
 })
 
@@ -205,7 +220,8 @@ test_that("can construct an id column", {
 
 test_that("vec_rbind() fails with arrays of dimensionality > 3", {
   expect_snapshot({
-    (expect_error(vec_rbind(array(NA, c(1, 1, 1))), "Can't bind arrays"))
+    (expect_error(vec_rbind(array(NA, c(1, 1, 1)))))
+    (expect_error(vec_rbind(array(NA, c(1, 1, 1)), .call = call("foo"))))
   })
 })
 
@@ -380,7 +396,10 @@ test_that("performance: Row binding with df-cols doesn't duplicate on every assi
 test_that("vec_cbind() reports error context", {
   expect_snapshot({
     (expect_error(vec_cbind(foobar(list()))))
+    (expect_error(vec_cbind(foobar(list()), .call = call("foo"))))
+
     (expect_error(vec_cbind(a = 1:2, b = int())))
+    (expect_error(vec_cbind(a = 1:2, b = int(), .call = call("foo"))))
   })
 })
 
@@ -475,8 +494,11 @@ test_that("can repair names in `vec_cbind()` (#227)", {
 })
 
 test_that("can supply `.names_to` to `vec_rbind()` (#229)", {
-  expect_error(vec_rbind(.names_to = letters), "must be")
-  expect_error(vec_rbind(.names_to = 10), "must be")
+  expect_snapshot({
+    (expect_error(vec_rbind(.names_to = letters)))
+    (expect_error(vec_rbind(.names_to = 10)))
+    (expect_error(vec_rbind(.names_to = letters, .call = call("foo"))))
+  })
 
   x <- data_frame(foo = 1:2, bar = 3:4)
   y <- data_frame(foo = 5L, bar = 6L)
@@ -572,8 +594,12 @@ test_that("names are not repaired if packed", {
 
 test_that("vec_cbind() fails with arrays of dimensionality > 3", {
   a <- array(NA, c(1, 1, 1))
-  expect_error(vec_cbind(a), "Can't bind arrays")
-  expect_error(vec_cbind(x = a), "Can't bind arrays")
+
+  expect_snapshot({
+    (expect_error(vec_cbind(a)))
+    (expect_error(vec_cbind(a, .call = call("foo"))))
+    (expect_error(vec_cbind(x = a)))
+  })
 })
 
 test_that("monitoring: name repair while cbinding doesn't modify in place", {
@@ -720,10 +746,16 @@ test_that("rbind repairs names of data frames (#704)", {
   expect_identical(vec_rbind(df), df_repaired)
   expect_identical(vec_rbind(df, df), vec_rbind(df_repaired, df_repaired))
 
-  expect_error(
-    vec_rbind(df, df, .name_repair = "check_unique"),
-    class = "vctrs_error_names_must_be_unique"
-  )
+  expect_snapshot({
+    (expect_error(
+      vec_rbind(df, df, .name_repair = "check_unique"),
+      class = "vctrs_error_names_must_be_unique"
+    ))
+    (expect_error(
+      vec_rbind(df, df, .name_repair = "check_unique", .call = call("foo")),
+      class = "vctrs_error_names_must_be_unique"
+    ))
+  })
 })
 
 test_that("vec_rbind() works with simple homogeneous foreign S3 classes", {
@@ -957,11 +989,13 @@ test_that("can't zap names when `.names_to` is supplied", {
     vec_rbind(foo = c(x = 1), .names_to = zap(), .name_spec = zap()),
     data.frame(x = 1)
   )
+
   expect_snapshot({
     (expect_error(
-      vec_rbind(foo = c(x = 1), .names_to = "id", .name_spec = zap()),
-      "Can't zap outer names when `.names_to` is supplied.",
-      fixed = TRUE
+      vec_rbind(foo = c(x = 1), .names_to = "id", .name_spec = zap())
+    ))
+    (expect_error(
+      vec_rbind(foo = c(x = 1), .names_to = "id", .name_spec = zap(), .call = call("foo"))
     ))
   })
 })