+++ NEWS.md
@@ -1,5 +1,7 @@
 # vctrs (development version)
 
+* `vec_rbind()` and `vec_cbind()` now have `.call` arguments (#1597).
+
 * `df_list()` has gained a new `.unpack` argument to optionally disable data
   frame unpacking (#1616).